requirements.txt
asciimatics==1.10.0
future==0.16.0
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.3.0
pyfiglet==0.7.5
wcwidth==0.1.7
